Reorganize signal editor code into some submodules
